Pulled out results content to new component
