+++ package.json
@@ -21,7 +21,8 @@
     "react-scripts": "^3.3.0",
     "redux": "^4.0.4",
     "redux-actions": "^2.6.5",
-    "redux-saga": "^1.0.5"
+    "redux-saga": "^1.0.5",
+    "@material-ui/icons": "latest"
   },
   "scripts": {
     "start": "react-scripts start",
