+++ src/components/widgets/UploadStepper.js
@@ -1,5 +1,5 @@
 /** @jsx jsx */
-import {useEffect, useState} from 'react';
+import {useState} from 'react';
 import { jsx } from '@emotion/core';
 import Stepper from "@material-ui/core/Stepper";
 import Step from "@material-ui/core/Step";
@@ -11,16 +11,7 @@ import SelectFile from "./SelectFile";
 import Button from "@material-ui/core/Button";
 import DuplicateActionRadioButtons from "./DuplicateActionRadioButtons";
 import UploadButton from "./UploadButton";
-import {useSelector} from "react-redux";
-import LeftSpinner from "./LeftSpinner";
-import Grid from "@material-ui/core/Grid";
-import List from "@material-ui/core/List";
-import ListItem from "@material-ui/core/ListItem";
-import ListItemIcon from "@material-ui/core/ListItemIcon";
-import ListItemText from "@material-ui/core/ListItemText";
-import CheckCircleOutlineIcon from '@material-ui/icons/CheckCircleOutline';
-import ErrorOutlineIcon from '@material-ui/icons/ErrorOutline';
-import HighlightOffIcon from '@material-ui/icons/HighlightOff';
+import UploadResultsContent from "./UploadResultsContent";
 
 const useStyles = makeStyles(theme => ({
   root: {
@@ -49,13 +40,9 @@ function getSteps() {
 export default function UploadStepper() {
   const classes = useStyles();
 
-  const sourceUploadStatus = useSelector(state => state.resources.sourceUploadStatus);
-  const sourceUploadData = useSelector(state => state.resources.sourceUploadData);
-
   const [activeStep, setActiveStep] = useState(0);
   const [selectedFile, setSelectedFile] = useState(null);
   const [selectedDuplicate, setSelectedDuplicate] = useState('ignore');
-  const [stepThreeContent, setStepThreeContent] = useState(<LeftSpinner />);
   const steps = getSteps();
 
   const handleBack = () => {
@@ -86,7 +73,7 @@ export default function UploadStepper() {
           handleNext={handleNext}
         />;
       case 3:
-        return stepThreeContent;
+        return <UploadResultsContent />;
     }
   }
   const backButton = () => {
@@ -116,77 +103,6 @@ export default function UploadStepper() {
       </Button>
     )
   };
-
-  useEffect(() => {
-    if (sourceUploadStatus === 'success') {
-      const failedFixityMessage = sourceUploadData.failed_fixity.length > 0
-          ? [
-            `The following files failed fixity: ${sourceUploadData.failed_fixity.join(', ')}`,
-            <ErrorOutlineIcon color="error"/>
-          ]
-          : [
-            'No files failed fixity',
-            <CheckCircleOutlineIcon color="primary"/>
-          ];
-      const resourcesIgnoredMessage = sourceUploadData.resources_ignored.length > 0
-        ? [
-          `The following duplicate resources were ignored: 
-          ${sourceUploadData.resources_ignored.join(', ')}`,
-          <ErrorOutlineIcon color="error"/>
-        ]
-        : [
-          'No duplicate resources were ignored',
-          <CheckCircleOutlineIcon color="primary"/>
-        ];
-      const resourcesUpdatedMessage = sourceUploadData.resources_updated.length > 0
-        ? [
-          `The following duplicate resources were updated:
-          ${sourceUploadData.resources_updated.join(', ')}`,
-          <ErrorOutlineIcon color="error"/>
-        ]
-        : [
-          'No duplicate resources were updated',
-          <CheckCircleOutlineIcon color="primary"/>
-        ];
-
-      const successfulMessage =
-        <Grid item md={6}>
-          <div>
-            <List dense={true}>
-              <ListItem>
-                <ListItemIcon>
-                  {failedFixityMessage[1]}
-                </ListItemIcon>
-                <ListItemText
-                  primary={failedFixityMessage[0]}
-                />
-              </ListItem>
-              <ListItem>
-                <ListItemIcon>
-                  {resourcesIgnoredMessage[1]}
-                </ListItemIcon>
-                <ListItemText
-                  primary={resourcesIgnoredMessage[0]}
-                />
-              </ListItem>
-              <ListItem>
-                <ListItemIcon>
-                  {resourcesUpdatedMessage[1]}
-                </ListItemIcon>
-                <ListItemText
-                  primary={resourcesUpdatedMessage[0]}
-                />
-              </ListItem>
-            </List>
-          </div>
-        </Grid>;
-
-      setStepThreeContent(successfulMessage);
-    }
-    else if (sourceUploadStatus === 'failure') {
-      setStepThreeContent(<div>{sourceUploadData.message}</div>);
-    }
-  }, [sourceUploadStatus]);
 
   return (
     <div className={classes.root}>
